Fixed components was placed in the wrong folder
