+++ Frontend/src/components/forms/CompanyForm.jsx
@@ -260,11 +260,7 @@ export default function CompanyForm({
         open={openModal}
         closeAfterTransition
         // submit on Enter key
-        onKeyDown={(e) => {
-          if (e.key === "Enter") {
-            submit();
-          }
-        }}
+
         // close on Escape key
         onClose={() => {
           setOpenModal(false);
