--- Frontend/src/components/forms/Form.css
@@ -1,30 +1,0 @@
-.FormModal .Box {
-  position: absolute;
-  top: 50%;
-  left: 50%;
-  transform: translate(-50%, -50%);
-
-  max-width: 98%;
-  width: 30rem;
-
-  max-height: 95%;
-  overflow-y: auto;
-
-  border-radius: min(3vw, 1.5rem);
-  padding: min(2vw, 1rem) min(3vw, 1.5rem);
-
-  background-color: whitesmoke;
-  box-shadow: #666 2px 2px 8px;
-}
-
-.FormModal h2 {
-  margin: 0;
-  text-transform: uppercase;
-}
-
-.FormModal .action-btns {
-  margin-block: 3%;
-
-  display: flex;
-  justify-content: space-between;
-}
