+++ Frontend/src/components/pages/partial/TableComponent.jsx
@@ -18,6 +18,7 @@ import {
   MoneyOff as MoneyOffIcon,
   Repeat as RepeatIcon,
   Email as EmailIcon,
+  AssignmentTurnedIn as AssignmentTurnedInIcon,
   Call as CallIcon,
   Work as WorkIcon,
   Error as ErrorIcon,
@@ -143,6 +144,9 @@ export default function TableComponent({
         .join(" ");
 
       switch (value) {
+        case "TODO":
+          icon = <AssignmentTurnedInIcon {...iconAttributes} />;
+          break;
         case "CONTACTED":
           icon = <CallIcon {...iconAttributes} />;
           break;
