--- components/pages/Home.jsx
@@ -1,11 +1,0 @@
-import { Typography } from "@mui/material";
-
-export default function Home() {
-  return (
-    <>
-      <Typography variant="h2" align="center" gutterBottom>
-        Home
-      </Typography>
-    </>
-  );
-}
