--- components/pages/NotFound.css
@@ -1,13 +1,0 @@
-.NotFound {
-  display: grid;
-  place-items: center;
-
-  text-align: center;
-}
-
-.NotFound > img {
-  max-width: 100%;
-  max-height: 100%;
-
-  padding-block: min(8vw, 4rem);
-}
