--- components/pages/pages.css
@@ -1,16 +1,0 @@
-.gumbi {
-    display: grid;
-    grid-template-columns: 1fr 1fr 1fr 1fr 1fr;
-    column-gap: 20px;
-    row-gap: 40px;
-    margin: 20px;
-}
-
-.project-topic{
-    color:rgb(12, 88, 189);
-    padding-left: 20px;
-}
-
-ul {
-    list-style-type: square;
-}
